Reorganize files
Create a src/pages directory for pages, and move other files to src/
