+++ packages/internal-playwright-helpers/tsconfig.json
@@ -7,7 +7,7 @@
     "lib": ["es2017", "dom"],
     "declaration": true,
     "outDir": "dist",
-    "rootDir": "./",
+    "rootDir": "./src",
     "strict": true,
     "esModuleInterop": true,
     // Prevent developers on different OSes from running into problems:
@@ -16,5 +16,5 @@
     "stripInternal": true,
     "allowJs": false
   },
-  "include": ["index.ts"]
+  "include": ["src/index.ts"]
 }
